requirements.txt
Flask==1.1.1
gunicorn==19.9.0
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
Flask-Script==2.0.6
SQLAlchemy==1.3.8
PyJWT==1.4.2
bcrypt==3.1.7
numpy==1.18.0
APScheduler==3.0.0
psycopg2==2.7.5
